clustering/requirements.txt
cycler==0.10.0
joblib==0.14.1
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.2
Pillow==7.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.45.0
